Friday morning

+++ publisher.py
@@ -1,108 +1,139 @@
-def lock(path):
-	import fcntl
-	x = open('path', 'w+')
-	fcntl.flock(x, fcntl.LOCK_EX | fcntl.LOCK_NB)
-	return x;
+def publish_queue(file_,path):
 
-def unlock(x):
-	import fcntl
-	fcntl.flock(x, fcntl.LOCK_UN);
+	print('File is put on the queue');
 
-def publish_queue(host, port, file_):
-	host = 'localhost';
-	import pika
-	import sys
-	#ssl_option = {'certfile': '/home/anilam/client/cert.pem', 'keyfile': '/home/anilam/client/key.pem'}
-	parameters = pika.ConnectionParameters(host=host, port=port);#
-	connection = pika.BlockingConnection(parameters)
+	host = '10.0.0.1';	
 
-	channel = connection.channel()
-	channel.confirm_delivery();
-
-	channel.queue_declare(queue='task_queue', durable=True)
-	channel.queue_declare(queue='result', durable = True);
-		
-
-	channel.basic_publish(exchange='',
-                      routing_key='task_queue',
-                      body=file_,
-                      properties=pika.BasicProperties(
-                         delivery_mode = 2, # make message persistent
-                      ))
-	def callback(ch, method, properties, body):
-    		print(" [x] Received response %r" % body)
-    		ch.basic_ack(delivery_tag = method.delivery_tag)
-
-	channel.basic_consume(callback,  queue='result')
-	channel.start_consuming()
-
-def unlock(x):
-	import fcntl
-	fcntl.flock(x, fcntl.LOCK_UN);
-
-def publish_queue(file_):
-	print('queue');
-	host = 'localhost';	
 	port = 5672
+
 	import pika
+
 	import sys
+
 	#ssl_option = {'certfile': '/home/anilam/client/cert.pem', 'keyfile': '/home/anilam/client/key.pem'}
+
 	parameters = pika.ConnectionParameters(host=host, port=port);#
+
 	connection = pika.BlockingConnection(parameters)
 
+
+
 	channel = connection.channel()
+
 	channel.confirm_delivery();
 
+
+
 	channel.queue_declare(queue='task_queue', durable=True)
+
 	channel.queue_declare(queue='result', durable = True);
+
 		
 
+
+
 	channel.basic_publish(exchange='',
+
                       routing_key='task_queue',
+
                       body=file_,
+
                       properties=pika.BasicProperties(
+
                          delivery_mode = 2, # make message persistent
+
                       ))
-	print("[X] Send " + file_);
+
+	print("[X] Send task:" + file_);
+
 	def callback(ch, method, properties, body):
-    		print(" [x] Received response %r" % body)
-		
+
+    		print(" [x] Received response from worker")
+
+		import os;
+
+		file_ = open(path+'.backup','w');
+
+		import flock;
+
+		fcntl.flock(file_, fcntl.LOCK_EX | fcntl.LOCK_NB);
+
+		file_.write(body);
+
+		fcntl.flock(file_,fcntl.LOCK_UN);
+
+		from shutil import copyfile
+
+		copyfile(path+'.backup', path);
+
+		os.remove(path+'.backup');	
+
     		ch.basic_ack(delivery_tag = method.delivery_tag)
+
 	channel.basic_consume(callback,  queue='result')
+
 	channel.start_consuming()
+
+
 
 def publish():
+
 	import time;
+
 	import os;
-	from threading import Thread
+
 	mypath = "/home/anilam/scripts/"
-	while 1:				
-		try:
+
+	while 1:	
+
+				
+
+		try:	
+
 			from os import listdir
+
 			from os.path import isfile, join
+
 			filelist = [f for f in listdir(mypath) if isfile(join(mypath, f))]
+
 			for f in filelist:
-				if (f.split('.')[-1]  != 'backup'):
-					import uuid
-					import os
-					from shutil import copyfile
-					x =  mypath+f+'.'+'backup';
+
+				if (f.split('.')[-1]  != 'result' and f.split('.')[-1]  != 'backup'):
+
+					print('New File to work with : ' + f);
+
 					filescr = mypath + f;
-					copyfile(filescr, x);			
-					file_ = open(filescr);
+
+					file_ = open(filescr,'r');
+
+					import fcntl;
+
+					fcntl.flock(file_, fcntl.LOCK_EX | fcntl.LOCK_NB);
+
 					text = file_.read();
-					desc = lock(filescr);
+
+					fcntl.flock(file_,fcntl.LOCK_UN);
+
 					file_.close();
-					unlock(desc);
+
 					os.remove(filescr);							
-					host = 'localhost';
+
+					host = '10.0.0.1';
+
 					port = 5672;
-					thread = Thread(target = publish_queue, args = (text, ))
-					thread.start()
-					
+
+					publish_queue(text, mypath + filescr+'.result');
+
+							
+
 		except:	
-			print('Waiting file to appear');
+
+			print('Error occured, waiting');
+
 			time.sleep(1);		
+
 			continue;
+
 	
+
 publish();
